backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-allauth>=0.61,<1.0
django-filter>=24.0,<25.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# AI Integration
anthropic>=0.18,<1.0

# Cloud Storage
django-storages[boto3]>=1.14,<2.0
boto3>=1.34,<2.0

# Utilities
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
PyPDF2>=3.0,<4.0
PyMuPDF>=1.24,<2.0

# Development
black>=24.0
flake8>=7.0
